Move styles into a seperate file for App
